updated response type

+++ src/main/java/com/taxicalls/billing/repository/BillingRepository.java
@@ -3,6 +3,6 @@ package com.taxicalls.billing.repository;
 import com.taxicalls.billing.model.Billing;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BillingRepository extends CrudRepository<Billing, Integer> {
+public interface BillingRepository extends CrudRepository<Billing, Long> {
 
 }
